+++ test/utest-online/testlinkapi_generic_online_test.py
@@ -474,6 +474,19 @@ class TestLinkAPIGenericOnlineTestCase(unittest.TestCase):
             self.client.getExecutionSet(
                             '40000713 plan', testcaseexternalid = 'TC-40000712')
          
+    def test_getRequirements_unknownID(self):
+        with self.assertRaisesRegex(TLResponseError, '7000.*40000712'):
+            self.client.getRequirements(
+                            '40000712 project', 
+                            testplanid = '40000713 plan', 
+                            platformid = '40000714 platform')
+         
+    def test_getReqCoverage_unknownID(self):
+        with self.assertRaisesRegex(TLResponseError, '7000.*40000712'):
+            self.client.getReqCoverage(
+                            '40000712 project', 
+                            '40000721 req')
+         
                                  
 if __name__ == "__main__":
     #import sys;sys.argv = ['', 'Test.testName']